requirements.txt
aiohttp>=3.8.1
pyyaml>=6.0.0
matplotlib>=3.10.0
markdown>=3.4.0